feat(drafts): ✨ Add Commissioner Predictions feature and permissions
Introduces a comprehensive "Commissioner Predictions" system to the Drafts module, including new domain models, repositories, services, and API endpoints for managing prediction seasons, contestants, rules, sets, entries, surrogates, results, standings, and carryovers. Adds new database tables and migrations, robust validation, error handling, and OpenAPI documentation. Updates permissions and role assignments to support managing, submitting, and reading predictions. Includes new integration and unit tests, infrastructure updates, and supporting assets.

+++ src/screendrafts.api/src/modules/drafts/ScreenDrafts.Modules.Drafts.Domain/Predictions/Entities/PredictionStanding.cs
@@ -24,7 +24,7 @@ public sealed class PredictionStanding : Entity<PredictionStandingId>
   public PredictionContestant Contestant { get; private set; } = default!;
   public ContestantId ContestantId { get; private set; } = default!;
 
-  public decimal Points { get; private set; } = default!;
+  public int Points { get; private set; } = default!;
   public DateTime? FirstCrossedTargetAtUtc { get; private set; } = default!;
   // methods: Add(points, targetPoints, beforeTotal, now)
 
@@ -37,7 +37,7 @@ public sealed class PredictionStanding : Entity<PredictionStandingId>
     return new PredictionStanding(season, contestant);
   }
 
-  public void Add(decimal points, int targetPoints, decimal beforeTotal, DateTime now)
+  public void Add(int points, int targetPoints, int beforeTotal, DateTime now)
   {
     Points += points;
     if (beforeTotal < targetPoints && Points >= targetPoints)
